FE/Qt: bugref:9973: Ups..
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@88249 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VirtualBox/src/runtime/UIMachineView.cpp
@@ -399,8 +399,8 @@ void UIMachineView::sltHandleActionTriggerViewScreenToggle(int iScreen, bool fEn
          * Do not send a hint if nothing has changed to prevent the guest being notified about its own changes. */
         if (   !isFullscreenOrSeamless()
             && uisession()->isGuestSupportsGraphics()
-            && (   m_pFrameBuffer->width() != (int)uWidth
-                || m_pFrameBuffer->height() != (int)uHeight
+            && (   m_pFrameBuffer->width() != uWidth
+                || m_pFrameBuffer->height() != uHeight
                 || uisession()->isScreenVisible(screenId()) != uisession()->isScreenVisibleHostDesires(screenId())))
             storeGuestSizeHint(QSize(uWidth, uHeight));
